requirements.txt
pillow==6.2.0
Django==2.2.4
djangorestframework==3.10.3
dj_database_url==0.5.0
psycopg2==2.8.4
django-rest-registration==0.5.2
django-extensions==2.2.5
boto3==1.10.34
django-storages==1.8